+++ TidalPy/RadialSolver/interfaces/reversed.pyx
@@ -7,12 +7,18 @@ cdef void cf_top_to_bottom_interface_bc(
         double complex* constant_vector_ptr,
         double complex* layer_above_constant_vector_ptr,
         double complex* uppermost_y_per_solution_ptr,
-        double gravity_upper, double layer_above_lower_gravity,
-        double density_upper, double layer_above_lower_density,
-        bool_cpp_t layer_is_solid, bool_cpp_t layer_above_is_solid,
-        bool_cpp_t layer_is_static, bool_cpp_t layer_above_is_static,
-        bool_cpp_t layer_is_incomp, bool_cpp_t layer_above_is_incomp,
-        unsigned char num_sols, unsigned char max_num_y
+        double gravity_upper,
+        double layer_above_lower_gravity,
+        double density_upper,
+        double layer_above_lower_density,
+        int layer_type,
+        int layer_above_type,
+        bint layer_is_static,
+        bint layer_above_is_static,
+        bint layer_is_incomp,
+        bint layer_above_is_incomp,
+        unsigned char num_sols,
+        unsigned char max_num_y
         ) noexcept nogil:
     
     # Loop variables
@@ -25,14 +31,14 @@ cdef void cf_top_to_bottom_interface_bc(
     cdef double liquid_density_at_interface
     interface_gravity = 0.5 * (gravity_upper + layer_above_lower_gravity)
     liquid_density_at_interface = NAN
-    if not layer_is_solid:
+    if not (layer_type == 0):
         if layer_is_static:
             liquid_density_at_interface = density_upper
-        elif not layer_above_is_solid and layer_above_is_static:
+        elif not (layer_above_type == 0) and layer_above_is_static:
             liquid_density_at_interface = layer_above_lower_density
         else:
             liquid_density_at_interface = density_upper
-    elif not layer_above_is_solid:
+    elif not (layer_above_type == 0):
         liquid_density_at_interface = layer_above_lower_density
 
     # Solve for constant vector
@@ -42,8 +48,9 @@ cdef void cf_top_to_bottom_interface_bc(
     cdef double complex lower_s1y1, lower_s1y2, lower_s1y5, lower_s1y6
     cdef double complex lower_s2y1, lower_s2y2, lower_s2y5, lower_s2y6
 
-    if layer_is_solid:
-        if layer_above_is_solid:
+    if (layer_type == 0):
+        # Solid Layer
+        if (layer_above_type == 0):
             # Both layers are solid. Constants are the same.
             solution_i = 0
             while num_sols > solution_i:
@@ -117,7 +124,7 @@ cdef void cf_top_to_bottom_interface_bc(
                 constant_vector_ptr[2] = y4_frac_1 * constant_vector_ptr[0] + y4_frac_2 * constant_vector_ptr[1]
     else:
         if layer_is_static:
-            if not layer_above_is_solid:
+            if not (layer_above_type == 0):
                 # Liquid layer above
                 if layer_above_is_static:
                     # Both layers are static liquids. Constants are the same.
@@ -135,7 +142,7 @@ cdef void cf_top_to_bottom_interface_bc(
                 #  layer above.
                 constant_vector_ptr[0] = layer_above_constant_vector_ptr[0]
         else:
-            if not layer_above_is_solid:
+            if not (layer_above_type == 0):
                 # Liquid layer above
                 if layer_above_is_static:
                     # Need to find 2 liquid (dynamic) constants from 1 liquid (static) constant