+++ TidalPy/RadialSolver/solutions.pyx
@@ -1,17 +1,18 @@
-# distutils: language = c++
+# distutils: language = c
 # cython: boundscheck=False, wraparound=False, nonecheck=False, cdivision=True, initializedcheck=False
 
 cdef size_t cf_find_num_solutions(
-        bool_cpp_t is_solid,
-        bool_cpp_t is_static,
-        bool_cpp_t is_incompressible
+        int layer_type,
+        bint is_static,
+        bint is_incompressible
         ) noexcept nogil:
     """ Determine number of solutions required for layer based on assumptions.
     
     Parameters
     ----------
-    is_solid : bool
-        Layer is solid (True) or liquid (False).
+    layer_type : int
+        - 0: Layer is solid 
+        - 1: Layer is liquid
     is_static : bool
         Use static (True) or dynamic (False) assumption.
     is_incompressible : bool
@@ -28,7 +29,8 @@ cdef size_t cf_find_num_solutions(
     cdef Py_ssize_t num_sols
     num_sols = 0
 
-    if is_solid:
+    if (layer_type == 0):
+        # Solid
         if is_static:
             if is_incompressible:
                 # TODO: Confirm
@@ -61,9 +63,9 @@ cdef size_t cf_find_num_solutions(
 
 
 def find_num_solutions(
-        bool_cpp_t is_solid,
-        bool_cpp_t is_static,
-        bool_cpp_t is_incompressible
+        int layer_type,
+        bint is_static,
+        bint is_incompressible
         ):
     
-    return cf_find_num_solutions(is_solid, is_static, is_incompressible)
+    return cf_find_num_solutions(layer_type, is_static, is_incompressible)