+++ TidalPy/defaultc.py
@@ -24,7 +24,7 @@ default_config_str = """
     # Logging levels
     # Options: DEBUG, INFO, WARNING, ERROR
     file_level = "DEBUG"
-    console_level = "DEBUG"
+    console_level = "INFO"
     console_error_level = "WARNING"
 
     # Determine if log should be printed to consol or written to disk when run in a jupyter notebook.